+++ src/api/interfaces/chain.rs
@@ -12,20 +12,117 @@
 */
 
 use crate::{
-	api::{error::Error, Api, ApiResult, TransactionStatus},
-	rpc::{HandleSubscription, Subscribe},
-	utils, XtStatus,
+	api::{Api, ApiResult},
+	rpc::{Request, Subscribe},
+	FromHexString,
 };
 use ac_compose_macros::rpc_params;
 pub use ac_node_api::{events::EventDetails, StaticEvent};
-use ac_node_api::{DispatchError, Events};
 use ac_primitives::{ExtrinsicParams, FrameSystemConfig};
 use log::*;
 use serde::de::DeserializeOwned;
-use sp_core::storage::StorageChangeSet;
+use sp_core::Pair;
+use sp_runtime::{generic::SignedBlock, traits::GetRuntimeBlockType, MultiSignature};
 
-pub type TransactionSubscriptionFor<Client, Hash> =
-	<Client as Subscribe>::Subscription<TransactionStatus<Hash, Hash>>;
+pub trait GetHeader<Hash> {
+	type Header;
+
+	fn get_finalized_head(&self) -> ApiResult<Option<Hash>>;
+
+	fn get_header(&self, hash: Option<Hash>) -> ApiResult<Option<Self::Header>>;
+}
+
+impl<Signer, Client, Params, Runtime> GetHeader<Runtime::Hash>
+	for Api<Signer, Client, Params, Runtime>
+where
+	Client: Request,
+	Runtime: FrameSystemConfig,
+	Params: ExtrinsicParams<Runtime::Index, Runtime::Hash>,
+	Runtime::Header: DeserializeOwned,
+	Runtime::Hash: FromHexString,
+{
+	type Header = Runtime::Header;
+
+	fn get_finalized_head(&self) -> ApiResult<Option<Runtime::Hash>> {
+		let finalized_block_hash =
+			self.client().request("chain_getFinalizedHead", rpc_params![])?;
+		Ok(finalized_block_hash)
+	}
+
+	fn get_header(&self, hash: Option<Runtime::Hash>) -> ApiResult<Option<Runtime::Header>> {
+		let block_hash = self.client().request("chain_getHeader", rpc_params![hash])?;
+		Ok(block_hash)
+	}
+}
+
+pub trait GetBlock<Number, Hash> {
+	type Block;
+
+	fn get_block_hash(&self, number: Option<Number>) -> ApiResult<Option<Hash>>;
+
+	fn get_block(&self, hash: Option<Hash>) -> ApiResult<Option<Self::Block>>;
+
+	fn get_block_by_num(&self, number: Option<Number>) -> ApiResult<Option<Self::Block>>;
+
+	/// A signed block is a block with Justification ,i.e., a Grandpa finality proof.
+	/// The interval at which finality proofs are provided is set via the
+	/// the `GrandpaConfig.justification_period` in a node's service.rs.
+	/// The Justification may be None.
+	fn get_signed_block(&self, hash: Option<Hash>) -> ApiResult<Option<SignedBlock<Self::Block>>>;
+
+	fn get_signed_block_by_num(
+		&self,
+		number: Option<Number>,
+	) -> ApiResult<Option<SignedBlock<Self::Block>>>;
+}
+
+impl<Signer, Client, Params, Runtime> GetBlock<Runtime::BlockNumber, Runtime::Hash>
+	for Api<Signer, Client, Params, Runtime>
+where
+	Signer: Pair,
+	MultiSignature: From<Signer::Signature>,
+	Client: Request,
+	Runtime: FrameSystemConfig + GetRuntimeBlockType,
+	Params: ExtrinsicParams<Runtime::Index, Runtime::Hash>,
+	Runtime::RuntimeBlock: DeserializeOwned,
+	Runtime::Hash: FromHexString,
+{
+	type Block = Runtime::RuntimeBlock;
+
+	fn get_block_hash(
+		&self,
+		number: Option<Runtime::BlockNumber>,
+	) -> ApiResult<Option<Runtime::Hash>> {
+		let block_hash = self.client().request("chain_getBlockHash", rpc_params![number])?;
+		Ok(block_hash)
+	}
+
+	fn get_block(&self, hash: Option<Runtime::Hash>) -> ApiResult<Option<Self::Block>> {
+		Self::get_signed_block(self, hash).map(|sb_opt| sb_opt.map(|sb| sb.block))
+	}
+
+	fn get_block_by_num(
+		&self,
+		number: Option<Runtime::BlockNumber>,
+	) -> ApiResult<Option<Self::Block>> {
+		Self::get_signed_block_by_num(self, number).map(|sb_opt| sb_opt.map(|sb| sb.block))
+	}
+
+	fn get_signed_block(
+		&self,
+		hash: Option<Runtime::Hash>,
+	) -> ApiResult<Option<SignedBlock<Self::Block>>> {
+		let block = self.client().request("chain_getBlock", rpc_params![hash])?;
+		Ok(block)
+	}
+
+	fn get_signed_block_by_num(
+		&self,
+		number: Option<Runtime::BlockNumber>,
+	) -> ApiResult<Option<SignedBlock<Self::Block>>> {
+		self.get_block_hash(number).map(|h| self.get_signed_block(h))?
+	}
+}
 
 pub trait ChainSubscription<Client, Hash>
 where
@@ -37,7 +134,7 @@ where
 	fn subscribe_finalized_heads(&self) -> ApiResult<Client::Subscription<Self::Header>>;
 }
 
-impl<Signer, Client, Params, Runtime> NodeSubscription<Client, Runtime::Hash>
+impl<Signer, Client, Params, Runtime> ChainSubscription<Client, Runtime::Hash>
 	for Api<Signer, Client, Params, Runtime>
 where
 	Client: Subscribe,
